+++ Content.Client/Ensnaring/EnsnareableSystem.cs
@@ -12,13 +12,14 @@ public sealed class EnsnareableSystem : SharedEnsnareableSystem
     {
         base.Initialize();
 
-        SubscribeLocalEvent<EnsnareableComponent, ComponentInit>(OnComponentInit);
         SubscribeLocalEvent<EnsnareableComponent, AppearanceChangeEvent>(OnAppearanceChange);
     }
 
-    private void OnComponentInit(EntityUid uid, EnsnareableComponent component, ComponentInit args)
+    protected override void OnEnsnareInit(Entity<EnsnareableComponent> ent, ref ComponentInit args)
     {
-        if(!TryComp<SpriteComponent>(uid, out var sprite))
+        base.OnEnsnareInit(ent, ref args);
+
+        if(!TryComp<SpriteComponent>(ent.Owner, out var sprite))
             return;
 
         // TODO remove this, this should just be in yaml.
